backend/farmconnect_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
httpx==0.24.1
supabase==2.0.0  # For storage uploads
requests==2.31.0  # For Nominatim geocode
python-multipart==0.0.6